Refactored grouping methods
Started a utility module for common types of grouping operations
