+++ materials_io/dft.py
@@ -1,5 +1,5 @@
 from typing import Union, Iterable, Tuple
-from materials_io.utils import preprocess_paths
+from materials_io.utils.grouping import preprocess_paths, group_by_postfix
 from materials_io.base import BaseParser
 from dfttopif import files_to_pif
 from operator import itemgetter
@@ -60,33 +60,8 @@ class DFTParser(BaseParser):
             ((files]): List of VASP files from the same calculation
         """
 
-        # Get the files that are likely VASP files, which we define as
-        #  those which start with the name of a known vasp file.
-        vasplike_files = []  # List of (path, type, (dir, postfix))
-        for filename in files:
-            # TODO (lw): This logic is likely useful elsewhere
-            # TODO (lw): We do not check if the files are from the same directory
-            # Find if the filename matches a known type
-            name = os.path.basename(filename)
-            name_lower = name.lower()
-            matches = [name_lower.startswith(n) for n in _vasp_file_names]
-            if not any(matches):
-                continue
-
-            # Get the extension of the file
-            match_id = matches.index(True)
-            vtype = _vasp_file_names[match_id]
-            ext = name[len(vtype):]
-            d = os.path.dirname(filename)
-
-            # Add to the list
-            vasplike_files.append((filename, vtype, (d, ext)))
-
-        # Group files by postfix type and directory
-        sort_key = itemgetter(2)
-        for k, group in itertools.groupby(sorted(vasplike_files, key=sort_key),
-                                          key=sort_key):
-            yield [x[0] for x in group]
+        for group in group_by_postfix(files, _vasp_file_names):
+            yield group
 
     def _group_pwscf(self, files: Iterable[str]) -> Iterable[Tuple[str, ...]]:
         """Assemble groups of files that are potentially PWSCF calculations
